+++ react/lib/JBLoading.tsx
@@ -4,6 +4,7 @@ import type {JBElementStandardProps} from 'jb-core/react'
 import 'jb-loading';
 import type {JBLoadingWebComponent} from 'jb-loading'
 import './module-declaration.js';
+import type { JBLoadingColor, JBLoadingSize } from './module-declaration.js';
 
 export const JBLoading = React.forwardRef((props:Props, ref) => {
   const element = React.useRef<HTMLElement>(null);
@@ -18,5 +19,5 @@ export const JBLoading = React.forwardRef((props:Props, ref) => {
   );
 });
 JBLoading.displayName = 'JBLoading';
-export type Props = PropsWithChildren<JBElementStandardProps<JBLoadingWebComponent>>
+export type Props = PropsWithChildren<JBElementStandardProps<JBLoadingWebComponent> & { size?: JBLoadingSize; color?: JBLoadingColor }>