+++ react/lib/module-declaration.ts
@@ -1,4 +1,5 @@
-export {};
+export type JBLoadingSize = "xs" | "sm" | "md" | "lg" | "xl";
+export type JBLoadingColor = "primary" | "secondary" | "neutral" | "content-inverse" | "content-primary" | "content-secondary";
 
 declare module "react" {
     namespace JSX {
@@ -7,6 +8,8 @@ declare module "react" {
       }
       interface JBLoadingType extends React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement>, HTMLElement> {
         "class"?:string,
+        size?: JBLoadingSize,
+        color?: JBLoadingColor,
       }
     }
 }